# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "EpiCare AI",
    "description": "\n**EpiCare AI** is a robust and comprehensive platform designed to manage and analyze epilepsy-related data, facilitating seamless interactions between patients, doctors, and caretakers.\n\n### Core Features\n- **Admin Dashboard**: Comprehensive system metrics and doctor PMDC verification tools.\n- **User Management**: Unified profiles tailored for Patients, Doctors, and Caretakers.\n- **Role-Based Authentication**: Secure access control with PMDC verification for doctors.\n- **Connection System**: Connect patients with verified medical professionals and trusted caretakers.\n- **AI-Driven EEG Analysis**: Upload EEG sessions, process signals, and receive detailed AI-generated seizure prediction reports.\n- **Personalized Dashboard**: Get real-time heuristic recommendations based on daily habits.\n- **Lifestyle & Medication Tracking**: Log sleep, stress, triggers, and medications with adherence algorithms.\n- **Emergency SOS**: Instantly dispatch multi-channel (Firebase/WhatsApp/Email) SOS alerts to configured contacts.\n\nThis API adheres strictly to REST principles, delivering standardized responses and explicit HTTP status codes to ensure a reliable developer experience.\n",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/emergency/contacts": {
      "get": {
        "tags": [
          "Emergency SOS"
        ],
        "summary": "List Emergency Contacts",
        "description": "Retrieves the list of configured emergency contacts for the authenticated patient.",
        "operationId": "get_emergency_contacts_api_v1_emergency_contacts_get",
        "responses": {
          "200": {
            "description": "A list of emergency contact objects.",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/EmergencyContactOut"
                  },
                  "type": "array",
                  "title": "Response Get Emergency Contacts Api V1 Emergency Contacts Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "Emergency SOS"
        ],
        "summary": "Add Emergency Contact",
        "description": "Adds a new emergency contact for the authenticated patient. A maximum of 3 emergency contacts are permitted per patient.",
        "operationId": "add_emergency_contact_api_v1_emergency_contacts_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmergencyContactCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "The newly created emergency contact object.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmergencyContactOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/emergency/contacts/{contact_id}": {
      "put": {
        "tags": [
          "Emergency SOS"
        ],
        "summary": "Update Emergency Contact",
        "description": "Updates an existing emergency contact's details (e.g., phone number or name).",
        "operationId": "update_emergency_contact_api_v1_emergency_contacts__contact_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "contact_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Contact Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmergencyContactUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmergencyContactOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Emergency SOS"
        ],
        "summary": "Delete Emergency Contact",
        "description": "Removes an emergency contact from the patient's list.",
        "operationId": "delete_emergency_contact_api_v1_emergency_contacts__contact_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "contact_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Contact Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/emergency/sos/trigger": {
      "post": {
        "tags": [
          "Emergency SOS"
        ],
        "summary": "Trigger Emergency SOS",
        "description": "Triggers an immediate Emergency SOS alert. This logs the event along with the patient's current GPS coordinates (if available) and instantly dispatches asynchronous background tasks to notify all registered emergency contacts via the configured provider (e.g., WhatsApp, Email).",
        "operationId": "trigger_sos_api_v1_emergency_sos_trigger_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SosTriggerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "An object acknowledging the trigger with the SOS event ID.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SosEventCreateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EmergencyContactCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 150,
            "minLength": 1,
            "title": "Name"
          },
          "relationship": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Relationship"
          },
          "phone_number": {
            "type": "string",
            "maxLength": 30,
            "minLength": 7,
            "pattern": "^\\+?[1-9]\\d{1,14}$",
            "title": "Phone Number"
          },
          "is_primary": {
            "type": "boolean",
            "title": "Is Primary",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "name",
          "relationship",
          "phone_number"
        ],
        "title": "EmergencyContactCreate",
        "description": "Request body for adding an emergency contact."
      },
      "EmergencyContactOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "relationship": {
            "type": "string",
            "title": "Relationship"
          },
          "phone_number": {
            "type": "string",
            "title": "Phone Number"
          },
          "is_primary": {
            "type": "boolean",
            "title": "Is Primary"
          },
          "verified": {
            "type": "boolean",
            "title": "Verified"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "name",
          "relationship",
          "phone_number",
          "is_primary",
          "verified",
          "created_at",
          "updated_at"
        ],
        "title": "EmergencyContactOut",
        "description": "Response model for an emergency contact."
      },
      "EmergencyContactUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "relationship": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Relationship"
          },
          "phone_number": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 30,
                "minLength": 7,
                "pattern": "^\\+?[1-9]\\d{1,14}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone Number"
          },
          "is_primary": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Primary"
          }
        },
        "type": "object",
        "title": "EmergencyContactUpdate",
        "description": "Request body for updating an emergency contact."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SosEventCreateResponse": {
        "properties": {
          "event_id": {
            "type": "integer",
            "title": "Event Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "event_id",
          "status",
          "message"
        ],
        "title": "SosEventCreateResponse",
        "description": "Confirmation response after triggering SOS."
      },
      "SosTriggerRequest": {
        "properties": {
          "latitude": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latitude"
          },
          "longitude": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Longitude"
          },
          "location_available": {
            "type": "boolean",
            "title": "Location Available",
            "default": false
          }
        },
        "type": "object",
        "title": "SosTriggerRequest",
        "description": "Request body for triggering an SOS alert."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}